+++ CHANGES.md
@@ -16,6 +16,12 @@ Since v2, ModTek adheres to [Semantic Versioning](http://semver.org/) for runtim
     - Use of ModTekSimpleInjector [ModTekSimpleInjector.Example.xml](ModTekSimpleInjector/ModTekSimpleInjector.Example.xml)
   - `Logs` section within `mod.json` -> probably unstable API (options and stuff seems to always change)
 
+## 4.5 - CptMoore
+
+For modders:
+- Updated HarmonyX to 2.15.0
+- Updated UnityDoorstop to 4.4.1
+
 ## 4.4 - CptMoore
 
 For modders: